requirements.txt
asgiref==3.11.0
boto3==1.42.15
botocore==1.42.15
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.2.9
django-allauth==65.13.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
jmespath==1.0.1
Markdown==3.10
oauthlib==3.3.1
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.16.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.2
wheel==0.45.1
gunicorn